Author data from the Milestones Project

Retrieve the author table containing information about authors associated with milestone events in the history of data visualization.

Usage

authors()

Format

A data frame. See authors() for details.

Source

https://datavis.ca/milestones/

Value

A data frame with 290 rows and 15 variables:

aid

author id, a numeric vector

prefix

name prefix, a character vector

givennames

given names, a character vector

lname

last name, a character vector

suffix

name suffix, a character vector

lived

years lived, a character vector

birthdate

birth date, a Date vector

birthplace

place of birth, a character vector

birthlat

latitude of birthplace, a numeric vector

birthlong

longitude of birthplace, a numeric vector

deathdate

death date, a Date vector

deathplace

place of death, a character vector

deathlat

latitude of deathplace, a numeric vector

deathlong

longitude of deathplace, a numeric vector

note

notes about this author, a character vector

Examples

if (FALSE) { # \dontrun{
# Get all authors
aut <- authors()

# Filter by birth location
library(dplyr)
aut |> filter(!is.na(birthplace))
} # }