Retrieve the mediaitem table from the milestones database

The mediaitem table gives each link or image used in displaying the milestones items

A dataset containing links and images used in displaying milestone items.

Usage

mediaitem()

mediaitem

Format

A data frame with 600 rows and 8 variables:

miid

mediaitem id, a numeric vector

type

type (link or image), a character vector

url

URL to the media item, a character vector

title

title or caption, a character vector

caption

caption text, a character vector

source

source of the media, a character vector

mid

milestone id to which this is attached, a numeric vector

type2

secondary type classification, a character vector

Source

https://datavis.ca/milestones/

Value

Returns a tibble of the mediaitem table

Details

The fields in the mediaitem table are:

miid

mediaitem id, a numeric vector

type

type, a character vector, with values link, image

url

URL to reference this item in the milestones application

title

a character vector

caption

a character vector

source

a character vector

mid

milestones id to which this is attached, a numeric vector

For type=="link", these give a standard HTML reference in the url field. For type=="image", the url field references the image in the milestones website.