Milestone data from the Milestones Project

Retrieve the milestone table containing milestone events in the history of data visualization. This is the primary table in the database. Other tables are linked to it via the mid key.

Usage

milestone()

Format

A data frame. See milestone() for details.

Source

https://datavis.ca/milestones/

Value

A data frame with 297 rows and 14 variables:

mid

milestone id, a numeric vector

slug

title slug, a character vector

date_from_numeric

numeric start date, a numeric vector

date_from

start date text, a character vector

date_to_numeric

numeric end date, a numeric vector

date_to

end date text, a character vector

tag

tag line, a character vector

description

description of the milestone, a character vector

location

location, a character vector

add_date

date added to database, a character vector

modified_date

date last modified, a character vector

status

publication status, a character vector

note

notes, a character vector

extra

extra information, a character vector

Examples

if (FALSE) { # \dontrun{
# Get all milestones
m <- milestone()

# Filter by date
library(dplyr)
m |> filter(date_from_numeric < 0)  # BC milestones
} # }