Retrieve the reference table from the milestones database

The reference table gives the references associated with milestones items

A dataset containing bibliographic references associated with milestone events.

Usage

reference()

reference

Format

A data frame with 352 rows and 17 variables:

rid

reference id, a numeric vector

type

reference type (article, book, incollection), a factor

author

author names, a character vector

title

title of publication, a character vector

journal

journal name (for articles), a character vector

month

month of publication, a character vector

year

year of publication, a character vector

volume

volume number, a character vector

number

issue number, a character vector

pages

page numbers, a character vector

publisher

publisher name, a character vector

address

publisher address, a character vector

editor

editor names, a character vector

booktitle

book title (for incollection), a character vector

bibtexkey

BibTeX key, a character vector

abstract

abstract, a character vector

note

notes, a character vector

Source

https://datavis.ca/milestones/

Value

Returns a tibble of the reference table

Details

The fields in the reference table are:

rid

reference id, a numeric vector

type

reference type, a factor with levels article book incollection

author

author, a character vector

title

title of publication, a character vector

journal

journal name, for an article, character vector

month

month name, a character vector

year

a numeric vector

volume

a character vector

number

a character vector

pages

a character vector

publisher

a character vector

address

a character vector

editor

a character vector

booktitle

a character vector

bibtexkey

BibTeX key used in the source .bib files, a character vector

abstract

a character vector

note

a character vector

In the note field, Loc{} refers to a library catalog identifier, e.g, Loc{BL: IA 3Q024}

TODO: need a function to turn these back to BibTeX